Avoid importing Gtk unnecessarily
Nobody wants to import Gtk in non-GUI programs just to check it's there.

Move Gtk-dependent classes into a separate module and put the user in
control. Don't play games with Gtk availability at all. Don't hide
whatever exception Gtk import might throw.
